# Fix OnEndOfDay not fired - Undeterministic ScheduledEvent triggering