--- a/Engine/RealTime/IRealTimeHandler.cs
+++ b/Engine/RealTime/IRealTimeHandler.cs
@@ -1,11 +1,11 @@
 ﻿/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,6 +16,7 @@
 
 using System;
 using System.ComponentModel.Composition;
+using QuantConnect.Data.UniverseSelection;
 using QuantConnect.Interfaces;
 using QuantConnect.Lean.Engine.Results;
 using QuantConnect.Packets;
@@ -46,7 +47,7 @@ namespace QuantConnect.Lean.Engine.RealTime
         /// Main entry point to scan and trigger the realtime events.
         /// </summary>
         void Run();
-        
+
         /// <summary>
         /// Set the current time for the event scanner (so we can use same code for backtesting and live events)
         /// </summary>
@@ -63,5 +64,10 @@ namespace QuantConnect.Lean.Engine.RealTime
         /// Trigger and exit signal to terminate real time event scanner.
         /// </summary>
         void Exit();
+
+        /// <summary>
+        /// Event fired each time that we add/remove securities from the data feed
+        /// </summary>
+        void OnSecuritiesChanged(SecurityChanges changes);
     }
 }