--- a/Tests/Engine/AlgorithmManagerTests.cs
+++ b/Tests/Engine/AlgorithmManagerTests.cs
@@ -307,6 +307,10 @@ namespace QuantConnect.Tests.Engine
             public void Exit()
             {
             }
+
+            public void OnSecuritiesChanged(SecurityChanges changes)
+            {
+            }
         }
 
         class NullTransactionHandler : ITransactionHandler