… Engine/RealTime/BaseRealTimeHandler.cs
… Engine/RealTime/LiveTradingRealTimeHandler.cs
… Tests/QuantConnect.Tests.csproj