<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Adding `IRealTimeHandler.OnSecuritiesChanged()` will be used to update
the `OnEndOfDay` security related scheduled events
- Adding `BaseRealTimeHandler.cs` to reduce code duplication in the
`Backtesting` and `LiveTrading` `RealTimeHandlers`
- Adding CSharp and Python regression tests
- Deprecating `OnEndOfDay()` callback because of two reasons, mainly
because Python does not support two methods with the same name, but also
because different assets have different market close times.
- `ScheduledEvents` set at the same time will now be deterministic. Adding unit test (_which fails in  master_)

**Performance tests StatefulCoarseUniverseSelectionBenchmark 2017 - 2018 - 250 securities (250 scheduled events added in this PR)**
`PR`
33.72 seconds at 57k data points per second. Processing total of 1,925,403 data points.
33.38 seconds at 58k data points per second. Processing total of 1,925,403 data points.
33.56 seconds at 57k data points per second. Processing total of 1,925,403 data points.
![imagen](https://user-images.githubusercontent.com/18473240/59918148-4299e600-93fa-11e9-976e-6bd807b258c7.png)

`master`
32.12 seconds at 60k data points per second. Processing total of 1,925,403 data points.
32.15 seconds at 60k data points per second. Processing total of 1,925,403 data points.
31.74 seconds at 61k data points per second. Processing total of 1,925,403 data points.

![imagen](https://user-images.githubusercontent.com/18473240/59918100-21d19080-93fa-11e9-8675-443434f59352.png)

**Note: there is a ~5% performance hit related to triggering the 250 security related `OnEndOfDay` scheduled events.**
#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/2625
Closes https://github.com/QuantConnect/Lean/issues/2865
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Fix bugs
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->