@jaredbroad 2019-06-21
Thank you Martin, we should make something which detects if the consumer uses the event handler. Few people do so it doesn't make sense to scan it each time.